Updated bot.cpp to check name of bot create before trying to create bot object
IsBotNameAvailable was a class function and a Bot object needed to be
created in order for it to work.  It turns out creating a bot object
crashes other users who have the same name.  So I made
IsBotNameAvailable a static function and pass the name to check into the
function.  This solves the crash problem.